Explicitly use global variables for settings
This is to enable support for fundle, a plugin manager for fish. Fundle
works by going through plugin directories and sourcing files itself,
and as this happens inside of a function set defaults to a function-
local variable; therefore the variables are not available at runtime
and the extension defaults to being disabled.